requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.2.0,<6.0
django-cors-headers>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
